Create Himmelblau’s function using machine learning
This code defines an input tensor x containing two elements, sets the requires_grad flag to True to enable automatic differentiation, and passes x as input to the himmelblau function to compute the function value func. The grad function is then called to compute the gradient of func with respect to x.

The grad function returns a tuple containing the gradient of func with respect to x, which is computed using the chain rule of differentiation. In this case, since x has two elements, the output gradient is also a tensor with two elements, corresponding to the partial derivatives of func with respect to x1 and x2.
So the output of the code is the gradient of the himmelblau function evaluated at the input point [1.0, -1.0].